start of restructure
