+++ gnss_lib_py/navdata/navdata.py
@@ -162,110 +162,6 @@ class NavData():
         for row_num in range(numpy_array.shape[0]):
             self[str(row_num)] = numpy_array[row_num,:]
 
-    def concat(self, navdata=None, axis=1, inplace=False):
-        """Concatenates second NavData instance by row or column.
-
-        Concatenates a second NavData instance to the existing NavData
-        instance by either row or column.
-
-        Each type of data is included in a row, so adding new rows with
-        ``axis=0``, means adding new types of data. Concat requires that
-        the new NavData matches the length of the existing NavData. Row
-        concatenation assumes the same ordering across both NavData
-        instances (e.g. sorted by timestamp) and does not perform any
-        matching/sorting itself.
-
-        You can also concatenate new columns ``axis=1``. If the row
-        names of the new NavData instance don't match the row names of
-        the existing NavData instance, the mismatched values will be
-        filled with np.nan.
-
-        Parameters
-        ----------
-        navdata : gnss_lib_py.parsers.navdata.NavData
-            Navdata instance to concatenate.
-        axis : int
-            Either add new rows (type) of data ``axis=0`` or new columns
-            (e.g. timesteps) of data ``axis=1``.
-        inplace : bool
-            If False, will return new concatenated NavData instance.
-            If True, will concatenate data to the current NavData
-            instance.
-
-        Returns
-        -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
-            If inplace is False, returns NavData instance after
-            concatenating specified data. If inplace is True, returns
-            None.
-
-        """
-
-        if not isinstance(navdata,NavData):
-            raise TypeError("concat input data must be a NavData instance.")
-
-        if axis == 0: # concatenate new rows
-            if len(self) != len(navdata):
-                raise RuntimeError("concat input data must be same " \
-                                 + "length to concatenate new rows.")
-            if not inplace:
-                new_navdata = self.copy()
-            for row in navdata.rows:
-                new_row_name = row
-                suffix = None
-                while new_row_name in self.rows:
-                    if suffix is None:
-                        suffix = 0
-                    else:
-                        suffix += 1
-                    new_row_name = row + "_" + str(suffix)
-                new_row = navdata[row].astype(navdata.orig_dtypes[row])
-                if inplace:
-                    self[new_row_name] = new_row
-                else:
-                    new_navdata[new_row_name] = new_row
-
-        elif axis == 1: # concatenate new columns
-            new_navdata = NavData()
-            # get unique list of row names
-            combined_rows = self.rows  + [row for row in navdata.rows
-                                          if row not in self.rows]
-
-            for row in combined_rows:
-                combined_row = np.array([])
-                # combine data from existing and new instance
-                for data in [self, navdata]:
-                    if row in data.rows:
-                        new_row = np.atleast_1d(data[row])
-                    elif len(data) == 0:
-                        continue
-                    else:
-                        # add np.nan for missing values
-                        new_row = np.empty((len(data),))
-                        new_row.fill(np.nan)
-                    new_row = np.array(new_row, ndmin=1)
-                    combined_row = np.concatenate((combined_row,
-                                                   new_row))
-                new_navdata[row] = combined_row
-
-            if len(self) > 0 and len(navdata) > 0:
-                new_navdata.orig_dtypes = navdata.orig_dtypes.copy()
-                new_navdata.orig_dtypes.update(self.orig_dtypes)
-            elif len(self) > 0:
-                new_navdata.orig_dtypes = self.orig_dtypes.copy()
-            elif len(navdata) > 0:
-                new_navdata.orig_dtypes = navdata.orig_dtypes.copy()
-
-            if inplace:
-                self.array = new_navdata.array
-                self.map = new_navdata.map
-                self.str_map = new_navdata.str_map
-                self.orig_dtypes = new_navdata.orig_dtypes.copy()
-
-        if inplace:
-            return None
-        return new_navdata
-
     def where(self, key_idx, value, condition="eq"):
         """Return NavData where conditions are met for the given row.
 
@@ -304,7 +200,7 @@ class NavData():
 
         Returns
         -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData
+        new_navdata : gnss_lib_py.navdata.navdata.NavData
             NavData with columns where given condition is satisfied
             for specified row
         """
@@ -426,93 +322,6 @@ class NavData():
         new_cols = np.squeeze(new_cols)
         return new_cols
 
-    def sort(self, order=None, ind=None, ascending=True,
-             inplace=False):
-        """Sort values along given row or using given index
-
-        Parameters
-        ----------
-        order : string/int
-            Key or index of the row on which NavData will be sorted
-        ind : list/np.ndarray
-            Ordering of indices to be used for sorting
-        ascending : bool
-            If true, sorts "ascending", otherwise sorts "descending"
-        inplace : bool
-            If False, will return new NavData instance with rows
-            sorted. If True, will sorted data rows in the
-            current NavData instance.
-
-        Returns
-        -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
-            If inplace is False, returns NavData instance after renaming
-            specified rows. If inplace is True, returns
-            None.
-
-        """
-        # check if there is only one column - no sorting needed
-        if self.shape[1] == 1:
-            return self
-
-        if ind is None:
-            assert order is not None, \
-            "Provide 'order' arg as row on which NavData is sorted"
-            if ascending:
-                ind = np.argsort(self[order])
-            else:
-                ind = np.argsort(-self[order])
-
-        if not inplace:
-            new_navdata = self.copy()   # create copy to return
-        for row_idx in range(self.shape[0]):
-            if inplace:
-                self.array[row_idx,:] = self.array[row_idx,ind]
-            else:
-                new_navdata.array[row_idx,:] = new_navdata.array[row_idx,ind]
-
-        if inplace:
-            return None
-        return new_navdata
-
-    def loop_time(self, time_row, delta_t_decimals=2):
-        """Generator object to loop over columns from same times.
-
-        Parameters
-        ----------
-        time_row : string/int
-            Key or index of the row in which times are stored.
-        delta_t_decimals : int
-            Decimal places after which times are considered equal.
-
-        Yields
-        ------
-        timestamp : float
-            Current timestamp.
-        delta_t : float
-            Difference between current time and previous time.
-        new_navdata : gnss_lib_py.parsers.navdata.NavData
-            NavData with same time, up to given decimal tolerance.
-
-        """
-
-        times = self[time_row]
-        times_unique = np.sort(np.unique(np.around(times,
-                                         decimals=delta_t_decimals)))
-        for time_idx, time in enumerate(times_unique):
-            if time_idx==0:
-                delta_t = 0
-            else:
-                delta_t = time-times_unique[time_idx-1]
-            new_navdata = self.where(time_row, [time-10**(-delta_t_decimals),
-                                                time+10**(-delta_t_decimals)],
-                                                condition="between")
-            if len(np.unique(new_navdata[time_row]))==1:
-                frame_time = new_navdata[time_row, 0]
-            else:
-                frame_time = time
-            yield frame_time, delta_t, new_navdata
-
     def is_str(self, row_name):
         """Check whether a row contained string values.
 
@@ -553,7 +362,7 @@ class NavData():
 
         Returns
         -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
+        new_navdata : gnss_lib_py.navdata.navdata.NavData or None
             If inplace is False, returns NavData instance after renaming
             specified rows. If inplace is True, returns
             None.
@@ -611,7 +420,7 @@ class NavData():
 
         Returns
         -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
+        new_navdata : gnss_lib_py.navdata.navdata.NavData or None
             If inplace is False, returns NavData instance after
             replacing specified data. If inplace is True, returns None.
 
@@ -671,7 +480,7 @@ class NavData():
 
         Returns
         -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData
+        new_navdata : gnss_lib_py.navdata.navdata.NavData
             Copy of original NavData with desired rows and columns
         """
         new_navdata = NavData()
@@ -710,7 +519,7 @@ class NavData():
 
         Returns
         -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
+        new_navdata : gnss_lib_py.navdata.navdata.NavData or None
             If inplace is False, returns NavData instance after removing
             specified rows and columns. If inplace is True, returns
             None.
@@ -767,64 +576,6 @@ class NavData():
             new_navdata[key] = new_row
         return new_navdata
 
-    def interpolate(self, x_row, y_rows, inplace=False, *args):
-        """Interpolate NaN values based on row data.
-
-        Additional ``*args`` arguments are passed into the ``np.interp``
-        function.
-
-        Parameters
-        ----------
-        x_row : string
-            Row name for x-coordinate of all values (e.g. gps_millis).
-            Row must not contain any nan values.
-        y_rows : list or string
-            Row name(s) for y-coordinate which includes nan values that
-            will be interpolated.
-        inplace : bool
-            If False, will return new NavData instance with nan values
-            interpolated. If True, will interpolate nan values within
-            the current NavData instance.
-
-        Returns
-        -------
-        new_navdata : gnss_lib_py.parsers.navdata.NavData or None
-            If inplace is False, returns NavData instance after removing
-            specified rows and columns. If inplace is True, returns
-            None.
-
-        """
-
-        if isinstance(y_rows,str):
-            y_rows = [y_rows]
-        if not isinstance(x_row, str):
-            raise TypeError("'x_row' must be row name as a string.")
-        if not isinstance(y_rows, list):
-            raise TypeError("'y_rows' must be single or list of " \
-                          + "row names as a string.")
-        self.in_rows([x_row] + y_rows)
-
-        if not inplace:
-            new_navdata = self.copy()
-        for y_row in y_rows:
-            nan_idxs = self.argwhere(y_row,np.nan)
-            if nan_idxs.size == 0:
-                continue
-            not_nan_idxs = self.argwhere(y_row,np.nan,"neq")
-            x_vals = self[x_row,nan_idxs]
-            xp_vals = self[x_row,not_nan_idxs]
-            yp_vals = self[y_row,not_nan_idxs]
-
-            if inplace:
-                self[y_row,nan_idxs] = np.interp(x_vals, xp_vals,
-                                                 yp_vals, *args)
-            else:
-                new_navdata[y_row,nan_idxs] = np.interp(x_vals, xp_vals,
-                                                        yp_vals, *args)
-        if inplace:
-            return None
-        return new_navdata
-
     def in_rows(self, rows):
         """Checks whether the given rows are in NavData.
 
@@ -851,104 +602,6 @@ class NavData():
         if len(missing_rows) > 0:
             raise KeyError(", ".join(missing_rows) + " row(s) are" \
                            + " missing from NavData object.")
-
-    def find_wildcard_indexes(self, wildcards, max_allow = None,
-                              excludes = None):
-        """Searches for indexes matching wildcard search input.
-
-        For example, a search for ``x_*_m`` would find ``x_rx_m`` or
-        ``x_sv_m`` or ``x_alpha_beta_gamma_m`` depending on the rows
-        existing in the NavData instance.
-
-        The ``excludes`` variable allows you to exclude indexes when
-        trying to match a wildcard. For example, if there are rows named
-        ``pr_raw_m``and ``pr_raw_sigma_m`` then the input
-        ``wildcards="pr_*_m", excludes=None`` would return
-        ``{"pr_*_m", ["pr_raw_m","pr_raw_sigma_m"]}`` but with the excludes
-        parameter set, the input ``wildcards="pr_*_m", excludes="pr_*_sigma_m"``
-        would only return ``{"pr_*_m", ["pr_raw_m"]}``
-
-        Will return an error no index is found matching the wildcard or
-        if more than ``max_allow`` indexes are found.
-
-        Currently only allows for a single wildcard '*' per index.
-
-        Parameters
-        ----------
-        wildcards : array-like or str
-            List/tuple/np.ndarray/set of indexes for which to search.
-        max_allow : int or None
-            Maximum number of valid indexes to allow before throwing an
-            error. If None, then no limit is placed.
-        excludes : array-like or str
-            List or string to exclude for each wildcard in wildcards.
-            Must be the same length as wildcards. Allowed to include a
-            wildcard '*' character but not necessary.
-
-        Returns
-        -------
-        wildcard_indexes : dict
-            Dictionary of the form {"search_term", [indexes,...]},
-
-        """
-
-        if isinstance(wildcards,str):
-            wildcards = [wildcards]
-        if not isinstance(wildcards, (list,tuple,np.ndarray,set)):
-            raise TypeError("wildcards input in find_wildcard_indexes" \
-                         +  " must be array-like or single string")
-        if not (isinstance(max_allow,int) or max_allow is None):
-            raise TypeError("max_allow input in find_wildcard_indexes" \
-                          + " must be an integer or None.")
-        # handle exclude types
-        if isinstance(excludes,str):
-            excludes = [excludes]
-        if excludes is None:
-            excludes = [None] * len(wildcards)
-        if not isinstance(excludes, (list,tuple,np.ndarray,set)):
-            raise TypeError("excludes input in find_wildcard_indexes" \
-                         +  " must be array-like, single string, " \
-                         + "or None for each wildcard")
-        if len(excludes) != len(wildcards):
-            raise TypeError("excludes input must match length of " \
-                          + "wildcard input.")
-        for ex_idx, exclude in enumerate(excludes):
-            if exclude is None or isinstance(exclude,str):
-                excludes[ex_idx] = [exclude]
-            if not isinstance(excludes[ex_idx], (list,tuple,np.ndarray,set)):
-                raise TypeError("excludes input in find_wildcard_indexes" \
-                             +  " must be array-like, single string, " \
-                             + "or None for each wildcard")
-
-        wildcard_indexes = {}
-
-        for wild_idx, wildcard in enumerate(wildcards):
-            if not isinstance(wildcard,str):
-                raise TypeError("wildcards must be strings")
-            if wildcard.count("*") != 1:
-                raise RuntimeError("One wildcard '*' and only one "\
-                          + "wildcard must be present in search string")
-            indexes = [row for row in self.rows
-                   if row.startswith(wildcard.split("*",maxsplit=1)[0])
-                    and row.endswith(wildcard.split("*",maxsplit=1)[1])]
-            if excludes[wild_idx] is not None:
-                for exclude in excludes[wild_idx]:
-                    if exclude is not None:
-                        if '*' in exclude:
-                            indexes = [row for row in indexes
-                                     if not (row.startswith(exclude.split("*",maxsplit=1)[0])
-                                     and row.endswith(exclude.split("*",maxsplit=1)[1]))]
-                        else:
-                            indexes = [row for row in indexes if exclude != row]
-            if max_allow is not None and len(indexes) > max_allow:
-                raise KeyError("More than " + str(max_allow) \
-                             + " possible row indexes for "  + wildcard)
-            if len(indexes) == 0:
-                raise KeyError("Missing " + wildcard + " row.")
-
-            wildcard_indexes[wildcard] = indexes
-
-        return wildcard_indexes
 
     def pandas_df(self):
         """Return pandas DataFrame equivalent to class
@@ -1201,7 +854,7 @@ class NavData():
 
         Returns
         -------
-        self: gnss_lib_py.parsers.NavData
+        self: gnss_lib_py.navdata.navdata
             Instantiation of NavData class with iteration initialized
         """
         self.curr_col = 0
@@ -1213,7 +866,7 @@ class NavData():
 
         Returns
         -------
-        x_curr : gnss_lib_py.parsers.NavData
+        x_curr : gnss_lib_py.navdata.navdata
             Current column (based on iteration count)
         """
         if self.curr_col >= self.num_cols:
