+++ tests/algorithms/test_fde.py
@@ -10,7 +10,7 @@ import os
 import pytest
 import numpy as np
 
-from gnss_lib_py.parsers.navdata import NavData
+from gnss_lib_py.navdata.navdata import NavData
 from gnss_lib_py.parsers.google_decimeter import AndroidDerived2022
 from gnss_lib_py.algorithms.fde import solve_fde, evaluate_fde
