+++ tests/algorithms/test_residuals.py
@@ -12,7 +12,7 @@ import numpy as np
 
 from gnss_lib_py.algorithms.snapshot import solve_wls
 from gnss_lib_py.parsers.google_decimeter import AndroidDerived2021
-from gnss_lib_py.parsers.navdata import NavData
+from gnss_lib_py.navdata.navdata import NavData
 from gnss_lib_py.algorithms.residuals import solve_residuals
 
 @pytest.fixture(name="root_path")
