+++ tests/algorithms/test_snapshot.py
@@ -12,7 +12,7 @@ import pytest
 import numpy as np
 
 from gnss_lib_py.parsers.google_decimeter import AndroidDerived2021, AndroidDerived2022
-from gnss_lib_py.parsers.navdata import NavData
+from gnss_lib_py.navdata.navdata import NavData
 from gnss_lib_py.algorithms.snapshot import wls, solve_wls
